32 bit yara and installer fixes

+++ source/RaccineLib/raccine.cpp
@@ -24,11 +24,10 @@
 #include "HandleWrapper.h"
 #include "Utils.h"
 #include "YaraRuleRunner.h"
-
+#include "EventLogHelper.h"
 #pragma comment(lib,"advapi32.lib")
 #pragma comment(lib,"shlwapi.lib")
 #pragma comment(lib,"Wbemuuid.lib")
-
 
 bool EvaluateYaraRules(const RaccineConfig& raccine_config,
     const std::wstring& lpCommandLine,
@@ -37,8 +36,16 @@ bool EvaluateYaraRules(const RaccineConfig& raccine_config,
     DWORD dwParentPid,
     DWORD dwGrandParentPid)
 {
+
+    std::wstring recent_event_details(L"\r\n\r\nRecentEvents:");
+
     if (raccine_config.is_debug_mode()) {
         wprintf(L"Running YARA on: %s\n", lpCommandLine.c_str());
+    }
+
+    if (raccine_config.use_eventlog_data_in_rules())
+    {
+        recent_event_details += eventloghelper::GetEvents();
     }
 
     WCHAR wTestFilename[MAX_PATH] = { 0 };
@@ -49,7 +56,7 @@ bool EvaluateYaraRules(const RaccineConfig& raccine_config,
     if (c == 0) {
         return false;
     }
-    utils::write_string_to_file(wTestFilename, lpCommandLine);
+    utils::write_string_to_file(wTestFilename, lpCommandLine + recent_event_details);
 
     BOOL fSuccess = TRUE;
 
@@ -86,7 +93,9 @@ bool EvaluateYaraRules(const RaccineConfig& raccine_config,
                 fRetVal = fProcessRetVal;
         }
     }
-    DeleteFileW(wTestFilename);
+    if (!raccine_config.is_debug_mode()) {
+        DeleteFileW(wTestFilename);
+    }
 
     // szYaraOutput comes formated with \0s to the end of the buffer
     outYaraOutput = outYaraOutput.substr(0, outYaraOutput.find(L'\0'));
@@ -131,156 +140,6 @@ void WriteEventLogEntryWithId(const std::wstring& pszMessage, DWORD dwEventId)
 void WriteEventLogEntry(const std::wstring& pszMessage)
 {
     WriteEventLogEntryWithId(pszMessage, RACCINE_DEFAULT_EVENTID);
-}
-
-bool is_malicious_command_line(const std::vector<std::wstring>& command_line)
-{
-    if (command_line.empty()) {
-        return false;
-    }
-
-    // Main programs to monitor
-    bool bVssadmin = false;
-    bool bWmic = false;
-    bool bWbadmin = false;
-    bool bcdEdit = false;
-    bool bPowerShell = false;
-    bool bDiskShadow = false;
-
-    // Command line params
-    bool bDelete = false;
-    bool bShadows = false;
-    bool bResize = false;
-    bool bShadowStorage = false;
-    bool bShadowCopy = false;
-    bool bCatalog = false;
-    bool bQuiet = false;
-    bool bRecoveryEnabled = false;
-    bool bIgnoreallFailures = false;
-    bool bWin32ShadowCopy = false;
-    const bool bEncodedCommand = does_command_line_contain_base64(command_line);
-    bool bVersion = false;
-
-    const std::wstring program = utils::getFileName(utils::to_lower(command_line[0]));
-    //wprintf(L"Program is: '%s'\n", program.c_str());
-    
-    // Check for invoked program
-    if (program == L"vssadmin.exe" || program == L"vssadmin") {
-        bVssadmin = true;
-    }
-
-    if (program == L"wmic.exe" || program == L"wmic") {
-        bWmic = true;
-    }
-
-    if (program == L"wbadmin.exe" || program == L"wbadmin") {
-        bWbadmin = true;
-    }
-
-    if (program == L"bcdedit.exe" || program == L"bcdedit") {
-        bcdEdit = true;
-    }
-
-    if (program == L"powershell.exe" || program == L"powershell") {
-        bPowerShell = true;
-    }
-
-    if (program == L"diskshadow.exe" || program == L"diskshadow") {
-        bDiskShadow = true;
-    }
-
-    // Check for keywords in command line parameters
-    std::vector<std::wstring> command_line_parameters(command_line.begin() + 1,
-        command_line.end());
-    for (const std::wstring& parameter : command_line_parameters) {
-        // Convert wchar to wide string so we can perform contains/find command
-        const std::wstring convertedArg(utils::to_lower(parameter));
-
-        // Simple flag checks
-        if (convertedArg == L"delete") {
-            bDelete = true;
-        }
-        else if (convertedArg == L"shadows") {
-            bShadows = true;
-        }
-        else if (convertedArg == L"shadowstorage") {
-            bShadowStorage = true;
-        }
-        else if (convertedArg == L"resize") {
-            bResize = true;
-        }
-        else if (convertedArg == L"shadowcopy") {
-            bShadowCopy = true;
-        }
-        else if (convertedArg == L"catalog") {
-            bCatalog = true;
-        }
-        else if (convertedArg == L"-quiet" || convertedArg == L"/quiet") {
-            bQuiet = true;
-        }
-        else if (convertedArg == L"recoveryenabled") {
-            bRecoveryEnabled = true;
-        }
-        else if (convertedArg == L"ignoreallfailures") {
-            bIgnoreallFailures = true;
-        }
-        else if (convertedArg.find(L"win32_shadowcopy") != std::string::npos) {
-            bWin32ShadowCopy = true;
-        }
-        else if (convertedArg == L"-version" || convertedArg == L"/version") {
-            bVersion = true;
-        }
-    }
-
-    // Check all combinations (our blocklist)
-    if ((bVssadmin && bDelete && bShadows) ||            // vssadmin.exe
-        (bVssadmin && bDelete && bShadowStorage) ||      // vssadmin.exe
-        (bVssadmin && bResize && bShadowStorage) ||      // vssadmin.exe
-        (bWmic && bDelete && bShadowCopy) ||             // wmic.exe
-        (bWbadmin && bDelete && bCatalog && bQuiet) || 	 // wbadmin.exe 
-        (bcdEdit && bIgnoreallFailures) ||               // bcdedit.exe
-        (bcdEdit && bRecoveryEnabled) ||                 // bcdedit.exe
-        (bPowerShell && bVersion) ||                     // powershell.exe
-        (bPowerShell && bWin32ShadowCopy) ||             // powershell.exe
-        (bPowerShell && bEncodedCommand) ||              // powershell.exe
-        (bDiskShadow && bDelete && bShadows)) {          // diskshadow.exe
-
-        // Activate blocking
-        return true;
-    }
-
-    return false;
-}
-
-bool does_command_line_contain_base64(const std::vector<std::wstring>& command_line)
-{
-    // Encoded Command List (Base64)
-    std::vector<std::wstring> encodedCommands = { L"JAB", L"SQBFAF", L"SQBuAH", L"SUVYI",
-                                                  L"cwBhA", L"aWV4I", L"aQBlAHgA",
-                                                  L"cwB", L"IAA", L"IAB", L"UwB" };
-
-    // Check for keywords in command line parameters
-    for (size_t iCount = 1; iCount < command_line.size(); iCount++) {
-
-        // Convert wchar to wide string so we can perform contains/find command
-        std::wstring convertedArg(utils::to_lower(command_line[iCount]));
-        std::wstring convertedArgOrig(command_line[iCount]);                        // original parameter (no tolower)
-        std::wstring convertedArgPrev(utils::to_lower(command_line[iCount - 1]));   // previous parameter
-
-        // Special comparison of current argument with previous argument
-        // allows to check for e.g. -encodedCommand JABbaTheHuttandotherBase64characters
-        if (convertedArgPrev.find(L"-e") == std::string::npos && convertedArgPrev.find(L"/e") == std::string::npos) {
-            continue;
-        }
-
-        for (const std::wstring& command : encodedCommands) {
-            if (convertedArgOrig.find(command) != std::string::npos) {
-                return true;
-            }
-        }
-    }
-
-    return false;
 }
 
 bool needs_powershell_workaround(const std::wstring& command_line)
